# V2 Auto Multiline Detection - Pattern Table Heuristic