--- a/pkg/logs/internal/decoder/auto_multiline_detection/json_detector.go
+++ b/pkg/logs/internal/decoder/auto_multiline_detection/json_detector.go
@@ -18,9 +18,9 @@ func NewJSONDetector() *JSONDetector {
 	return &JSONDetector{}
 }
 
-// Process checks if a message is a JSON message.
+// ProcessAndContinue checks if a message is a JSON message.
 // This implements the Herustic interface - so we should stop processing if we detect a JSON message by returning false.
-func (j *JSONDetector) Process(context *messageContext) bool {
+func (j *JSONDetector) ProcessAndContinue(context *messageContext) bool {
 	if jsonRegexp.Match(context.rawMessage) {
 		context.label = noAggregate
 		return false