--- a/pkg/logs/internal/decoder/auto_multiline_detection/pattern_table.go
+++ b/pkg/logs/internal/decoder/auto_multiline_detection/pattern_table.go
@@ -119,7 +119,7 @@ func (p *PatternTable) DumpTable() []DiagnosticRow {
 }
 
 // ProcessAndContinue adds a pattern to the table and updates its label based on it's frequency.
-// This implements the Herustic interfce - so we should stop processing if the label was changed
+// This implements the Herustic interface - so we should stop processing if the label was changed
 // due to pattern detection.
 func (p *PatternTable) ProcessAndContinue(context *messageContext) bool {