--- a/pkg/logs/internal/decoder/auto_multiline_detection/tokenizer.go
+++ b/pkg/logs/internal/decoder/auto_multiline_detection/tokenizer.go
@@ -345,5 +345,5 @@ func isMatch(seqA []tokens.Token, seqB []tokens.Token, thresh float64) bool {
 		}
 	}
 
-	return match >= requiredMatches
+	return true
 }