<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?

This PR introduces the pattern table heuristic for Auto Multiline Detection and aggregation. 
This is the last heuristic defined in the RFC for the initial version of this feature. 

The pattern table exists to assist in diagnostics and reporting telemetry. Telemetry and diagnostic reporting will come in the next PR. One of the pieces of diagnostics we want to see is the distribution of patterns in a log source which the pattern table stores for us. The pattern table may also be used in the future to apply labels for certain distribution of log patterns

### Implementation details

The pattern table makes the assumption that most log formats are rather consistent and have few unique prefix formats. However there is nothing stopping someone from tailing `/dev/urandom` which would produce consistently unique "formats". As a result, we have to ensure the table doesn't grow forever. 

Since we only care about the most popular formats, we can solve this by: 
- keeping the size limited
- keeping it sorted by frequency 
- evicting patterns we haven't seen in a while if the table becomes full

The eviction strategy is a simple LRU where  the index of the message is stored and if the table becomes full, the pattern with the least recently updated index is evicted. 

<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

This PR also optimizes the tokenizer `isMatch` function to exit early when a match is not met which improves performance in non matching cases, but retains the same impact in matching cases. 

```
before
BenchmarkTokenizerIsMatch-10    	  	 1000000	      1212 ns/op	       0 B/op	       0 allocs/op
BenchmarkTokenizerIsMatchNoMatchEnd-10    	 1296780	      1018 ns/op	       0 B/op	       0 allocs/op
BenchmarkTokenizerIsMatchFullMatch-10    	  998858	      1238 ns/op	       0 B/op	       0 allocs/op
after
BenchmarkTokenizerIsMatchNoMatch-10    	  	 3885801	      295.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkTokenizerIsMatchNoMatchEnd-10    	 1000000	      1014 ns/op	       0 B/op	       0 allocs/op
BenchmarkTokenizerIsMatchFullMatch-10    	  859585	      1227 ns/op	       0 B/op	       0 allocs/op

```

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

Here is an example of the kind of data the pattern table can report - something like this will be available on the status page in the future

```
445      start_group     *CCCCC'CCCCCCCCCC'-DDDD-DD-DD DD:DD:DD
102      aggregate       CC:'C'DDDCDDDDDDDDDDCDDDDDDDDCDCDCCDDDDDDDDDD'
74       aggregate       CCCCCC CCCCCCCCCC CCC CCCCCCCC :'CCCCCCCCCC
74       aggregate       CCC CC CC CCCCCCCCCC CCC CCCCCCCC :'CCCCCCCCCC
73       aggregate       CCCCCC CC CCCCCCCCC(CC:CC.CC):CCCCCCCC 'D:DD.DDDDDD' CCCCCC 
72       aggregate       CC:'C'DDDCDDDDDDDDDDCDDDDDDDCCCDDCCDDDDDDDDDD'
43       aggregate       CC:'C'DDDCDDDDDDDDDDCDDCDDDDCDCDDCCDDDDDDDDDD'
38       aggregate        CCCCCC CCCCCCC 
37       aggregate       CCCCCCCC CC CCCCC 
37       aggregate       CCCCCC CC CCCCCCCCC CCC CCCCC(CC:CC.CC): CCCCCCCC 'D:DD.DDD
6        aggregate       CC:'C'DDDCDDDDDDDDDDCCDDDDDDDDDDCCDDDDDDDDDD'
```

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

No Functional change